remove workflow.py
